Move file ids to own file
